requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
google-generativeai>=0.4.0
requests>=2.31.0
dnspython>=2.3.0
